requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
mlxtend==0.22.0
pyyaml==6.0.1
python-dotenv==1.0.0
# scikit-learn==1.3.0
# scipy==1.11.2